% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdmulti_package.R
\docType{package}
\name{rdmulti-package}
\alias{rdmulti-package}
\alias{_PACKAGE}
\alias{rdmulti_package}
\title{rdmulti: analysis of RD Designs with multiple cutoffs or scores}
\description{
The regression discontinuity (RD) design is a popular quasi-experimental design
for causal inference and policy evaluation. The \code{'rdmulti'} package provides tools
to analyze RD designs with multiple cutoffs or scores: \code{\link{rdmc}()} estimates
pooled and cutoff-speficif effects in multi-cutoff designs, \code{\link{rdmcplot}()}
draws RD plots for multi-cutoff RD designs and \code{\link{rdms}()} estimates effects in
cumulative cutoffs or multi-score designs. For more details, and related \code{Stata} and
\code{R} packages useful for analysis of RD designs, visit \url{https://sites.google.com/site/rdpackages}.
}
\references{
M.D. Cattaneo, R. Titiunik and G. Vazquez-Bare. (2020). \href{https://sites.google.com/site/rdpackages/rdmulti/Cattaneo-Titiunik-VazquezBare_2020_Stata.pdf}{Analysis of Regression Discontinuity Designs with Multiple Cutoffs or Multiple Scores}. \emph{Working paper}.
}
\author{
Matias Cattaneo, Princeton University. \email{cattaneo@princeton.edu}

Rocio Titiunik, Princeton University. \email{titiunik@princeton.edu}

Gonzalo Vazquez-Bare, UC Santa Barbara. \email{gvazquez@econ.ucsb.edu}
}
